Mash it all together because damn pika/pack
