+++ .babelrc
@@ -6,8 +6,5 @@
       }
     }]
   ],
-  "plugins": ["@babel/plugin-proposal-class-properties"],
-  "ignore": [
-    "**/*.jar"
-  ]
+  "plugins": ["@babel/plugin-proposal-class-properties"]
 }
