+++ package.json
@@ -33,24 +33,17 @@
       ],
       [
         "@pika/plugin-build-node"
-      ],
-      [
-        "@pika/plugin-copy-assets",
-        {
-          "files": [
-            "assets"
-          ]
-        }
       ]
     ]
   },
   "devDependencies": {
+    "@babel/cli": "^7.7.0",
     "@babel/core": "^7.4.3",
+    "@babel/node": "^7.7.0",
     "@babel/plugin-proposal-class-properties": "^7.4.0",
     "@babel/preset-env": "^7.5.4",
     "@pika/pack": "^0.5.0",
     "@pika/plugin-build-node": "^0.7.1",
-    "@pika/plugin-copy-assets": "^0.7.1",
     "@pika/plugin-standard-pkg": "^0.7.1",
     "babel-eslint": "^10.0.1",
     "babel-jest": "^24.9.0",
